+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add options to the CLI command to specify a host and port.
+  Useful for testing without the need to create a config file.
 ### Deprecated
 - Factory methods will become instance-based with the same name,
   in the next major version.